What I will be working on over the weekend
Updated backlog with what I will be working on over the weekend sense I
finished what I needed to get done during week one as well as what I
would like to work on during week two. One of the things I will be
working on is a document that shows everything that goes into front-end
development and why it has a lot of technical depth.